Set a lightweight schedular for collecting metrics every 10 seconds and persist metrics in db,the purpose is analyse metrics and create an average of them daily,weekly or monthly.

+++ src/database.py
@@ -1,4 +1,4 @@
-from sqlalchemy import MetaData, INTEGER, FLOAT
+from sqlalchemy import MetaData, INTEGER, FLOAT, BIGINT
 from sqlalchemy.orm import DeclarativeBase, MappedAsDataclass
 from sqlalchemy.ext.asyncio import (
     AsyncEngine,
@@ -18,8 +18,8 @@ class Base(DeclarativeBase, MappedAsDataclass):
         types.MetricId: INTEGER,
         types.MemoryUsage: FLOAT,
         types.CpuUsage: FLOAT,
-        types.DiskSpaceTotal: INTEGER,
-        types.DiskSpaceUsed: INTEGER,
+        types.DiskSpaceTotal: BIGINT,
+        types.DiskSpaceUsed: BIGINT,
         types.DiskUsage: FLOAT
     }
